timestamp to ms since 1971

+++ customtypes/logs.go
@@ -3,9 +3,9 @@ package customtypes
 // Log is a log with a container
 // name attached.
 type Log struct {
-	ContainerID   string  `json:"ContainerID,omitempty"`
-	ContainerName string  `json:"ContainerName,omitempty"`
-	Log           string  `json:"msg"`
-	ServiceName   string  `json:"sourceServiceName"`
-	LogTimestamp  float64 `json:"rt"`
+	ContainerID   string `json:"ContainerID,omitempty"`
+	ContainerName string `json:"ContainerName,omitempty"`
+	Log           string `json:"msg"`
+	ServiceName   string `json:"sourceServiceName"`
+	LogTimestamp  uint64 `json:"rt"`
 }
